requirements.txt
numpy
sklearn
pillow
tensorflow